Title: 최소 통과 시간, Time: 380ms, Memory: 24MB, Status: Wrong Answer - Codetree

+++ 250611/최소 통과 시간/minimum-transit-time.java
@@ -10,7 +10,7 @@ public class Main {
         
         int left = 1;
         int right = 1000000000;
-        int answer = Long.MAX_VALUE;
+        int answer = Integer.MAX_VALUE;
 
         while (left <= right) {
